javascript confirm before delete action

+++ admin/includes/view_all_posts.php
@@ -120,13 +120,15 @@
                     $output .= "<td>{$post_date}</td>";
                     $output .= "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
                     $output .= "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                    $output .= "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+                    $output .= "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
                     $output .= "</tr>";
 
                     echo $output;
                 }
 
             ?>
+
+
         </tbody>
     </table>
 </form>
